requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
httpx
python-multipart